Ran back to React
Was having trouble with Nuxt, it felt like it was hiding too much from me. Not sure if I added that in another commit but I'll add it here. Took the Tailwind next template and I'll be goin from here but my VS Code isn't autofilling so that's getting annoying. Kinda really relying on autocomplete :|